implements Navids request of integrating vocab and schema into API
