+++ src/cript/api/schema.py
@@ -1,41 +1,9 @@
-import json
+from typing import Union
 
-from jsonschema import validate
+from api import API, _get_global_cached_api
 
 
-class CRIPTSchema:
-    """
-    Data class representing the node schema of CRIPT.
-    """
-
-    _data: dict = {}
-
-    def __init__(self, db_schema: dict):
-        # TODO check that we support that version
-
-        self._data = db_schema
-
-    def is_node_valid(self, node_json: str) -> bool:
-        """
-        checks a node JSON schema against the db schema to return if it is valid or not.
-        This function does not take into consideration vocabulary validation.
-        For vocabulary validation please check `is_vocab_valid`
-
-        Parameters
-        ----------
-        node:
-            a node in JSON form
-
-        Returns
-        -------
-        bool
-            whether the node JSON is valid or not
-        """
-
-        # TODO currently validate says every syntactically valid JSON is valid
-        # TODO do we want invalid schema to raise an exception?
-        node_dict = json.loads(node_json)
-        if validate(node_dict, self._data):
-            return True
-        else:
-            return False
+def is_node_valid(node_json: str, api: Union[API, None] = None) -> bool:
+    if api is None:
+        api = _get_global_cached_api()
+    return api.is_node_valid(node_json)
